+++ app/main.py
@@ -4,7 +4,7 @@ from fastapi import FastAPI, Depends, Body
 from fastapi.middleware.cors import CORSMiddleware
 from contextlib import asynccontextmanager
 
-from app.routers import users, watchlist, admin
+from app.routers import users, watchlist, admin, professors
 from app.alerts.scheduler import continuous_check
 from app.utils.logger import get_logger
 from app.utils.constants import ALLOWED_ORIGINS
@@ -36,3 +36,4 @@ app.add_middleware(
 app.include_router(admin.router)
 app.include_router(users.router)
 app.include_router(watchlist.router)
+app.include_router(professors.router)
